+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 *   Much better reporting of parse and conversion errors.
 *   Support for custom and remote types.
 *   Support for receive with timeout.
+*   ifdef/ifndef/undef didn't handle capitalized macro names. Fixed.
 *   Catch assumed the kind was an atom and didn't accept an expression. Fixed.
 
 ## v0.0.1 (2015-12-28)